Moves all append vec "meta" types into submodule
