+++ accounts-db/src/append_vec.rs
@@ -4,12 +4,16 @@
 //!
 //! <https://docs.solanalabs.com/implemented-proposals/persistent-account-storage>
 
+#[cfg(feature = "dev-context-only-utils")]
+pub(crate) mod meta;
+#[cfg(not(feature = "dev-context-only-utils"))]
+mod meta;
+
+pub mod test_utils;
+
 use {
     crate::{
-        account_storage::{
-            meta::{AccountMeta, StoredMeta},
-            stored_account_info::{StoredAccountInfo, StoredAccountInfoWithoutData},
-        },
+        account_storage::stored_account_info::{StoredAccountInfo, StoredAccountInfoWithoutData},
         accounts_file::{
             AccountsFileError, InternalsForArchive, MatchAccountOwnerError, Result, StorageAccess,
             StoredAccountsInfo, ALIGN_BOUNDARY_OFFSET,
@@ -23,6 +27,7 @@ use {
     },
     log::*,
     memmap2::MmapMut,
+    meta::{AccountMeta, StoredAccountNoData, StoredMeta},
     solana_account::{AccountSharedData, ReadableAccount, WritableAccount},
     solana_clock::Epoch,
     solana_hash::Hash,
@@ -44,7 +49,8 @@ use {
     thiserror::Error,
 };
 
-pub mod test_utils;
+pub(crate) use meta::StoredAccountMeta;
+pub use meta::StoredMetaWriteVersion;
 #[cfg(test)]
 use solana_account::accounts_equal;
 
@@ -117,159 +123,6 @@ impl<'a> ValidSlice<'a> {
     #[cfg(all(unix, test))]
     pub(crate) fn slice(&self) -> &[u8] {
         self.0
-    }
-}
-
-/// References to account data stored elsewhere. Getting an `Account` requires cloning
-/// (see `StoredAccountMeta::clone_account()`).
-#[derive(PartialEq, Eq, Debug)]
-pub struct StoredAccountMeta<'append_vec> {
-    pub meta: &'append_vec StoredMeta,
-    /// account data
-    pub account_meta: &'append_vec AccountMeta,
-    pub(crate) data: &'append_vec [u8],
-    pub(crate) offset: usize,
-    pub(crate) stored_size: usize,
-    pub(crate) hash: &'append_vec AccountHash,
-}
-
-impl<'append_vec> StoredAccountMeta<'append_vec> {
-    pub fn pubkey(&self) -> &'append_vec Pubkey {
-        &self.meta.pubkey
-    }
-
-    pub fn hash(&self) -> &'append_vec AccountHash {
-        self.hash
-    }
-
-    pub fn stored_size(&self) -> usize {
-        self.stored_size
-    }
-
-    pub fn offset(&self) -> usize {
-        self.offset
-    }
-
-    pub fn data(&self) -> &'append_vec [u8] {
-        self.data
-    }
-
-    pub fn data_len(&self) -> usize {
-        self.meta.data_len as usize
-    }
-
-    pub fn meta(&self) -> &StoredMeta {
-        self.meta
-    }
-}
-
-/// [`StoredAccountMeta`] without account data or account hash.
-struct StoredAccountNoData<'append_vec> {
-    meta: &'append_vec StoredMeta,
-    account_meta: &'append_vec AccountMeta,
-    offset: usize,
-    stored_size: usize,
-}
-
-impl<'append_vec> StoredAccountNoData<'append_vec> {
-    #[inline(always)]
-    fn lamports(&self) -> u64 {
-        self.account_meta.lamports
-    }
-
-    #[inline(always)]
-    fn owner(&self) -> &'append_vec Pubkey {
-        &self.account_meta.owner
-    }
-
-    #[inline(always)]
-    fn pubkey(&self) -> &'append_vec Pubkey {
-        &self.meta.pubkey
-    }
-
-    fn sanitize(&self) -> bool {
-        self.sanitize_executable() && self.sanitize_lamports()
-    }
-
-    #[inline(always)]
-    fn offset(&self) -> usize {
-        self.offset
-    }
-
-    #[inline(always)]
-    fn stored_size(&self) -> usize {
-        self.stored_size
-    }
-
-    #[inline(always)]
-    fn data_len(&self) -> u64 {
-        self.meta.data_len
-    }
-
-    #[inline(always)]
-    fn executable(&self) -> bool {
-        self.account_meta.executable
-    }
-
-    #[inline(always)]
-    fn rent_epoch(&self) -> Epoch {
-        self.account_meta.rent_epoch
-    }
-
-    fn sanitize_executable(&self) -> bool {
-        // Sanitize executable to ensure higher 7-bits are cleared correctly.
-        self.ref_executable_byte() & !1 == 0
-    }
-
-    /// Check if the account data matches that of a default account.
-    ///
-    /// Note that we are not comparing against AccountSharedData::default() because we do not have access to the account data,
-    /// so we compare data _length_ in lieu of actual data. This check otherwise identical to AccountSharedData::default().
-    fn is_default_account(&self) -> bool {
-        self.account_meta.lamports == 0
-            && self.meta.data_len == 0
-            && !self.account_meta.executable
-            && self.account_meta.rent_epoch == Epoch::default()
-            && self.account_meta.owner == Pubkey::default()
-    }
-
-    fn sanitize_lamports(&self) -> bool {
-        // Check if the account data matches that of a default account if it has 0 lamports.
-        self.account_meta.lamports != 0 || self.is_default_account()
-    }
-
-    fn ref_executable_byte(&self) -> &u8 {
-        // Use extra references to avoid value silently clamped to 1 (=true) and 0 (=false)
-        // Yes, this really happens; see test_new_from_file_crafted_executable
-        let executable_bool: &bool = &self.account_meta.executable;
-        let executable_bool_ptr = ptr::from_ref(executable_bool);
-        // UNSAFE: Force to interpret mmap-backed bool as u8 to really read the actual memory content
-        let executable_byte: &u8 = unsafe { &*(executable_bool_ptr.cast()) };
-        executable_byte
-    }
-}
-
-impl IsZeroLamport for StoredAccountNoData<'_> {
-    fn is_zero_lamport(&self) -> bool {
-        self.lamports() == 0
-    }
-}
-
-impl<'append_vec> ReadableAccount for StoredAccountMeta<'append_vec> {
-    fn lamports(&self) -> u64 {
-        self.account_meta.lamports
-    }
-    fn data(&self) -> &'append_vec [u8] {
-        self.data()
-    }
-    fn owner(&self) -> &'append_vec Pubkey {
-        &self.account_meta.owner
-    }
-    fn executable(&self) -> bool {
-        self.account_meta.executable
-    }
-    fn rent_epoch(&self) -> Epoch {
-        self.account_meta.rent_epoch
     }
 }
