+++ runtime/src/snapshot_package.rs
@@ -7,12 +7,12 @@ use {
     agave_feature_set as feature_set,
     log::*,
     solana_accounts_db::{
-        account_storage::meta::StoredMetaWriteVersion,
         accounts::Accounts,
         accounts_db::AccountStorageEntry,
         accounts_hash::{
             AccountsDeltaHash, AccountsHash, AccountsHashKind, MerkleOrLatticeAccountsHash,
         },
+        append_vec::StoredMetaWriteVersion,
         epoch_accounts_hash::EpochAccountsHash,
     },
     solana_sdk::{
